feat: added vue3 frontend; developed new backend api

+++ crates/airlab-lib/src/ctx/mod.rs
@@ -4,7 +4,7 @@ pub use self::error::{Error, Result};
 
 #[derive(Clone, Debug)]
 pub struct Ctx {
-    user_id: i32,
+    user_id: i64,
 }
 
 impl Ctx {
@@ -13,7 +13,7 @@ impl Ctx {
         Self { user_id: 0 }
     }
 
-    pub const fn new(user_id: i32) -> Result<Self> {
+    pub const fn new(user_id: i64) -> Result<Self> {
         if user_id == 0 {
             Err(Error::CtxCannotNewRootCtx)
         } else {
@@ -24,7 +24,7 @@ impl Ctx {
 
 impl Ctx {
     #[must_use]
-    pub const fn user_id(&self) -> i32 {
+    pub const fn user_id(&self) -> i64 {
         self.user_id
     }
 }
